% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_lines}
\alias{all_lines}
\title{Line samples of migration corridors}
\format{
A list of sf polylines for all species in the database
}
\usage{
all_lines
}
\description{
10,000 lines randomly generated inside of the migration corridors linking non-UK regions to UK regions
to sample the site for possible migrating birds
}
\keyword{datasets}
