% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_polygons}
\alias{all_polygons}
\title{Polygons of Migration corridors}
\format{
A list of sf polygons for all species in the database
}
\usage{
all_polygons
}
\description{
The polygons of the migration corridors for each species as per the BTO report. This is saved as a list and accessed
by way of a call to the object.
}
\keyword{datasets}
