% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{betaInputPars_densPlots}
\alias{betaInputPars_densPlots}
\title{function for Beta density plot of model input parameters}
\usage{
betaInputPars_densPlots(p, stdev, fill = "olivedrab", xlab)
}
\description{
function for Beta density plot of model input parameters
}
