% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{defaultSpeciesValues}
\alias{defaultSpeciesValues}
\title{Default parameters for all species}
\format{
A nested list of 25 species
\describe{
  \item{speciesname}{the name of the species, under which a nested list of parameters is found}
  ...
}
}
\usage{
defaultSpeciesValues
}
\description{
A list of all species and their parameters for loading into the app
}
\keyword{datasets}
