% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EMOD_OSWFs_WGS84_UK}
\alias{EMOD_OSWFs_WGS84_UK}
\title{UK Offshore wind farms}
\format{
A spatial polygon dataframe
\describe{
  \item{COUNTRY}{Country that wind farm is associated with (should be UK only)}
  \item{NAME}{The name of the OSWF}
  \item{N_TURBINES}{Number of planned or operational turbines at a site}
  \item{POWER_MW}{Approximate power production of site}
  \item{STATUS}{Legal status of site (Planned, Approved, Production, or Dismantled)}
  \item{YEAR}{Year windfarm went into production}
  \item{COAST_DIST}{Distance to coast of polygon centroid}
  \item{AREA_SQKM}{Total windfarm area in kilometers squared}
  \item{NOTES}{Additional notes}
  \item{Shape_Leng}{Longest length of the polygon}
  \item{Shape_Area}{Area in hectares}
}
}
\usage{
EMOD_OSWFs_WGS84_UK
}
\description{
A spatial polygon dataframe of all the UK Offshore windfarms from EMODNet accessed July 2021
https://www.emodnet-humanactivities.eu/search-results.php?dataname=Wind+Farms+%28Polygons%29
}
\keyword{datasets}
