% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{get_wf_width}
\alias{get_wf_width}
\title{Get windfarm width}
\usage{
get_wf_width(polyg)
}
\arguments{
\item{polyg}{A polygon. A polygon in WGS84 for calculating the distances between vertices.}
}
\value{
The maximum distance between polygon vertices
}
\description{
Returns the width of the windfarm in kilometers for use in the flux calculations.
}
