% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{getTempFolderName}
\alias{getTempFolderName}
\title{generate a temporary folder name for the current session to write out the simulation output files}
\usage{
getTempFolderName()
}
\description{
generate a temporary folder name for the current session to write out the simulation output files
}
