% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{label.help}
\alias{label.help}
\title{Function to create icons and labels with info about parameters}
\usage{
label.help(label, id)
}
\description{
Function to create icons and labels with info about parameters
}
