% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{normDens_ParsPlots}
\alias{normDens_ParsPlots}
\title{function for Normal density plot of model input parameters}
\usage{
normDens_ParsPlots(
  mu,
  stdev,
  fill = "olivedrab",
  xlab,
  refValue_E = mu,
  refValue_SD = stdev
)
}
\description{
function for Normal density plot of model input parameters
}
