% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{normDens_ParsQtls}
\alias{normDens_ParsQtls}
\title{function for Normal density quantiles of model input parameters}
\usage{
normDens_ParsQtls(mu, stdev, varTag, decPlaces = 3)
}
\description{
function for Normal density quantiles of model input parameters
}
