% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{NormNumericInput}
\alias{NormNumericInput}
\title{NormNumericInput}
\usage{
NormNumericInput(
  paramID,
  varName,
  infoId = "foo",
  infoText = "",
  E_value = 50,
  E_min = 1,
  E_max = 100,
  E_step = 1,
  SD_value = 50,
  SD_min = 1,
  SD_max = 100,
  SD_step = 1,
  via_InsertUI = FALSE,
  Button = TRUE,
  cellWidths = "30\%"
)
}
\description{
Creates a Numeric input with a SD input as well
}
