% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SCOTWIND_SHAPES}
\alias{SCOTWIND_SHAPES}
\title{Scotwind lease areas}
\format{
A spatial polygon dataframe
\describe{
  \item{COUNTRY}{Country that wind farm is associated with (should be UK only)}
  \item{NAME}{The name of the OSWF}
  \item{N_TURBINES}{Number of planned or operational turbines at a site}
  \item{POWER_MW}{Approximate power production of site}
  \item{STATUS}{Legal status of site (Planned, Approved, Production, or Dismantled)}
  \item{YEAR}{Year windfarm went into production}
  \item{COAST_DIST}{Distance to coast of polygon centroid}
  \item{AREA_SQKM}{Total windfarm area in kilometers squared}
  \item{NOTES}{Additional notes}
  \item{Shape_Leng}{Longest length of the polygon}
  \item{Shape_Area}{Area in hectares}
}
}
\usage{
SCOTWIND_SHAPES
}
\description{
A spatial polygon of Scotwind offshore lease areas
}
\keyword{datasets}
