% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{species_data}
\alias{species_data}
\title{Species data with wingspans and lengths}
\format{
A dataframe
\describe{
  \item{Voous}{Species Voous}
  \item{Common}{Common English name of species}
  \item{Scientific}{Scientific name of species}
  \item{Length}{Mean length of species (cm)}
  \item{Wingspan}{Mean wingspan of species (cm)}
}
}
\usage{
species_data
}
\description{
A dataframe with species names, wingspans and wing lengths
}
\keyword{datasets}
