% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_FHD}
\alias{template_FHD}
\title{Start up template for bootstrap outputs}
\format{
A dataframe with 201 columns 
\describe{
  \item{bootId_n,...}{bootstrapped height of a bird for nth bootstrap}
  ...
}
}
\usage{
template_FHD
}
\description{
A template for what bootstrap outputs should be formatted as
}
\keyword{datasets}
