% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{template_monthDens_samples}
\alias{template_monthDens_samples}
\title{Template for the bootstrapped monthly densities}
\format{
A dataframe with 12 variables (for months)
\describe{
  \item{monthname}{Rows for wind availability, mean downtime and sd of downtime}
  ...
}
}
\usage{
template_monthDens_samples
}
\description{
A template dataset
}
\keyword{datasets}
