% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_helpers.R
\name{truncNormPars_densPlots}
\alias{truncNormPars_densPlots}
\title{function for Truncated Normal density plot of model input parameters}
\usage{
truncNormPars_densPlots(
  mu,
  stdev,
  lower = -Inf,
  upper = Inf,
  fill = "olivedrab",
  xlab
)
}
\description{
function for Truncated Normal density plot of model input parameters
}
