% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{turbineOpTempTable}
\alias{turbineOpTempTable}
\title{Temporary template table for windfarm parameters by month}
\format{
A dataframe with 12 variables (for months)
\describe{
  \item{monthname}{Rows for wind availability, mean downtime and sd of downtime}
  ...
}
}
\usage{
turbineOpTempTable
}
\description{
Temporary template table for windfarm parameters by month
}
\keyword{datasets}
